requirements.txt
streamlit
pillow
pytesseract
opencv-python
numpy